extract modal
